backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
youtube-transcript-api
google-generativeai
pydantic
pydantic-settings
typer
